[#515] Basic syncing and rendering of our Panel render object for the smart glasses. This is still a prototype and would not support other types of objects (like texts, circles or whatever we plan for the future)

+++ src/main/java/de/srendi/advancedperipherals/common/smartglasses/modules/overlay/propertyTypes/FloatingNumberProperty.java
@@ -12,7 +12,7 @@ import java.lang.annotation.Target;
 @ObjectProperty(type = FloatingNumberType.class)
 public @interface FloatingNumberProperty {
 
-    double min() default Double.MIN_VALUE;
-    double max() default Double.MAX_VALUE;
+    float min() default Float.MIN_VALUE;
+    float max() default Float.MAX_VALUE;
 
 }
